requirements.txt
python-telegram-bot==20.1
pydub==0.25.1
GoogleBard==0.0.7
SpeechRecognition==3.10.0
boto3
ffmpeg